Added "Sticky" Posts
Sticky posts can now be used to be featured on the home page. Only one post can be featured or they will be stacked incorrectly. "Recent Posts" title is a dirty fix, looks fine for the end user but it actually forms part of of the sticky entry-header.

Sticky posts will only be sticky on the home page.